+++ src/compas_rhino/geometry/brep/brep.py
@@ -1,17 +1,16 @@
 from compas.geometry import Frame
 from compas.geometry import Brep
-from compas.geometry import BrepInvalidError
 from compas.geometry import BrepTrimmingError
 from compas.geometry import Plane
 
 from compas_rhino.conversions import box_to_rhino
-from compas_rhino.conversions import point_to_rhino
 from compas_rhino.conversions import xform_to_rhino
 from compas_rhino.conversions import frame_to_rhino
 from compas_rhino.conversions import cylinder_to_rhino
 
 import Rhino
 
+from .builder import RhinoBrepBuilder
 from .face import RhinoBrepFace
 from .edge import RhinoBrepEdge
 from .vertex import RhinoBrepVertex
@@ -36,6 +35,8 @@ class RhinoBrep(Brep):
         The list of vertex geometries as points in 3D space.
     edges : list[:class:`~compas_rhino.geometry.RhinoBrepEdge`], read-only
         The list of edges which comprise this brep.
+    trims : list[:class:`~compas_rhino.geometry.RhinoBrepTrim`], read-only
+        The list of trims which comprise this brep.
     loops : list[:class:`~compas_rhino.geometry.RhinoBrepLoop`], read-only
         The list of loops which comprise this brep.
     faces : list[:class:`~compas_rhino.geometry.RhinoBrepFace`], read-only
@@ -70,18 +71,22 @@ class RhinoBrep(Brep):
 
     @property
     def data(self):
-        faces = []
-        for face in self.faces:
-            faces.append(face.data)
-        return {"faces": faces}
+        return {
+            "vertices": [v.data for v in self.vertices],
+            "edges": [e.data for e in self.edges],
+            "faces": [f.data for f in self.faces],
+        }
 
     @data.setter
     def data(self, data):
-        faces = []
-        for facedata in data["faces"]:
-            face = RhinoBrepFace.from_data(facedata)
-            faces.append(face)
-        self._create_native_brep(faces)
+        builder = RhinoBrepBuilder()
+        for v_data in data["vertices"]:
+            RhinoBrepVertex.from_data(v_data, builder)
+        for e_data in data["edges"]:
+            RhinoBrepEdge.from_data(e_data, builder)
+        for f_data in data["faces"]:
+            RhinoBrepFace.from_data(f_data, builder)
+        self._brep = builder.result
 
     # ==============================================================================
     # Properties
@@ -102,6 +107,11 @@ class RhinoBrep(Brep):
 
     @property
     def edges(self):
+        if self._brep:
+            return [RhinoBrepEdge(edge) for edge in self._brep.Edges]
+
+    @property
+    def trims(self):
         if self._brep:
             return [RhinoBrepEdge(trim) for trim in self._brep.Trims]
 
@@ -326,88 +336,3 @@ class RhinoBrep(Brep):
         """
         resulting_breps = self._brep.Split(cutter.native_brep, TOLERANCE)
         return [RhinoBrep.from_native(brep) for brep in resulting_breps]
-
-    # ==============================================================================
-    # Other Methods
-    # ==============================================================================
-
-    def _create_native_brep(self, faces):
-        # Source: https://github.com/mcneel/rhino-developer-samples/blob/3179a8386a64602ee670cc832c77c561d1b0944b/rhinocommon/cs/SampleCsCommands/SampleCsTrimmedPlane.cs
-        # Things need to be defined in a valid brep:
-        #  1- Vertices
-        #  2- 3D Curves (geometry)
-        #  3- Edges (topology - reference curve geometry)
-        #  4- Surfaces (geometry)
-        #  5- Faces (topology - reference surface geometry)
-        #  6- Loops (2D parameter space of faces)
-        #  4- Trims and 2D curves (2D parameter space of edges)
-        self._brep = Rhino.Geometry.Brep()
-        for face in faces:
-            rhino_face, rhino_surface = self._create_brep_face(face)
-            for loop in face.loops:
-                rhino_loop = self._brep.Loops.Add(Rhino.Geometry.BrepLoopType.Outer, rhino_face)
-                for edge in loop.edges:
-                    start_vertex, end_vertex = self._add_edge_vertices(edge)
-                    rhino_edge = self._add_edge(edge, start_vertex, end_vertex)
-                    rhino_2d_curve = self._create_trim_curve(rhino_edge, rhino_surface)
-                    self._add_trim(rhino_2d_curve, rhino_edge, rhino_loop)
-
-        self._brep.Repair(TOLERANCE)
-        self._brep.JoinNakedEdges(
-            TOLERANCE
-        )  # without this, Brep.Trim() led to some weird results on de-serialized Breps
-        self._validate_brep()
-
-    def _validate_brep(self):
-        if self._brep.IsValid:
-            return
-
-        error_message = ""
-        valid_topo, log_topo = self._brep.IsValidTopology()
-        valid_tol, log_tol = self._brep.IsValidTolerancesAndFlags()
-        valid_geo, log_geo = self._brep.IsValidGeometry()
-        if not valid_geo:
-            error_message += "Invalid geometry:\n{}\n".format(log_geo)
-        if not valid_topo:
-            error_message += "Invalid topology:\n{}\n".format(log_topo)
-        if not valid_tol:
-            error_message += "Invalid tolerances:\n{}\n".format(log_tol)
-
-        raise BrepInvalidError(error_message)
-
-    def _create_brep_face(self, face):
-        # Geometry
-        surface_index = self._brep.AddSurface(face.native_surface)
-        brep_surface = self._brep.Surfaces.Item[surface_index]
-        # Topology
-        brep_face = self._brep.Faces.Add(surface_index)
-        return brep_face, brep_surface
-
-    def _add_edge_vertices(self, edge):
-        start_vertex = self._brep.Vertices.Add(point_to_rhino(edge.start_vertex.point), TOLERANCE)
-        end_vertex = self._brep.Vertices.Add(point_to_rhino(edge.end_vertex.point), TOLERANCE)
-        return start_vertex, end_vertex
-
-    def _add_edge(self, edge, start_vertex, end_vertex):
-        # Geometry
-        curve_index = self._brep.AddEdgeCurve(edge.curve)
-        # Topology
-        rhino_edge = self._brep.Edges.Add(start_vertex, end_vertex, curve_index, TOLERANCE)
-        return rhino_edge
-
-    def _add_trim(self, rhino_trim_curve, rhino_edge, rhino_loop):
-        # Geometry
-        trim_curve_index = self._brep.AddTrimCurve(rhino_trim_curve)
-        # Topology
-        trim = self._brep.Trims.Add(rhino_edge, True, rhino_loop, trim_curve_index)
-        trim.IsoStatus = getattr(
-            Rhino.Geometry.IsoStatus, "None"
-        )  # IsoStatus.None makes lint, IDE and even Python angry
-        trim.TrimType = Rhino.Geometry.BrepTrimType.Boundary
-        trim.SetTolerances(TOLERANCE, TOLERANCE)
-
-    @staticmethod
-    def _create_trim_curve(rhino_edge, rhino_surface):
-        curve_2d = rhino_surface.Pullback(rhino_edge.EdgeCurve, TOLERANCE)
-        curve_2d.Reverse()
-        return curve_2d